Revert "56 refactor the class gamestatus to fit the state pattern"
