--- src/main/java/com/group34/GameState/BetweenRoundState.java
@@ -1,18 +1,0 @@
-package com.group34.GameState;
-
-import com.group34.TowerDefence;
-
-public class BetweenRoundState implements GameState {
-    
-    public void enterState(TowerDefence towerDefence) {
-        // Display round number
-    }
-
-    public void exitState(TowerDefence towerDefence) {
-        // Go to next round
-    }
-
-    public void update(TowerDefence towerDefence) {
-        // Check if player wants to start next round
-    }
-}
