--- src/main/java/com/group34/GameState/GameOverState.java
@@ -1,19 +1,0 @@
-package com.group34.GameState;
-
-import com.group34.TowerDefence;
-
-public class GameOverState implements GameState {
-    
-    public void enterState(TowerDefence towerDefence) {
-        // Display game over screen
-    }
-
-    public void exitState(TowerDefence towerDefence) {
-        //Go to main menu
-    }
-
-    public void update(TowerDefence towerDefence) {
-        // Check if player wants to restart or go to main menu
-    }
-    
-}
