+++ src/main/java/com/group34/Main.java
@@ -2,10 +2,12 @@ package com.group34;
 
 import java.awt.Dimension;
 import java.awt.geom.Point2D;
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
-import com.group34.GameState.GameState;
-import com.group34.GameState.MidRoundState;
+import javax.swing.JFrame;
+
 import com.group34.Model.Board.Board;
 import com.group34.Model.Cash.CashVault;
 import com.group34.Model.Enemy.Enemy;
@@ -15,22 +17,26 @@ import com.group34.Model.Game.Game;
 import com.group34.Model.Game.Player;
 import com.group34.Model.Road.RoadBuilder;
 import com.group34.Model.Road.RoadSpawn;
+import com.group34.Model.Road.RoadToken;
 import com.group34.Model.Round.Round;
+import com.group34.Model.Round.RoundBuilder;
+import com.group34.Model.Round.RoundEvent;
 import com.group34.Model.Round.RoundConfig;
 import com.group34.Model.Tower.LightningSmurfFactory;
 import com.group34.Model.Tower.Tower;
+import com.group34.View.BoardView;
+import com.group34.View.Shop.ShopController;
 import com.group34.View.Shop.ShopModel;
 
 
 class TowerDefenceBuilder {
     Board board;
+    List<Round> rounds;
     CashVault cashVault;
     Game game;
     Player player;
     RoadSpawn roadSpawn;
     ShopModel shopModel;
-    List<Round> rounds;
-    GameState startState;
     GameSpeed gameSpeed;
 
     public TowerDefenceBuilder setBoard(Board board) {
@@ -45,6 +51,11 @@ class TowerDefenceBuilder {
 
     public TowerDefenceBuilder setPlayer(Player player) {
         this.player = player;
+        return this;
+    }
+
+    public TowerDefenceBuilder setRounds(List<Round> rounds) {
+        this.rounds = rounds;
         return this;
     }
 
@@ -63,35 +74,98 @@ class TowerDefenceBuilder {
         return this;
     }
 
-    public TowerDefenceBuilder setRounds (List<Round> rounds) {
-        this.rounds = rounds;
-        return this;
-    }
-
-    public TowerDefenceBuilder setStartState (GameState startState) {
-        this.startState = startState;
-        return this;
-    }
-
     public TowerDefence build() {
         return new TowerDefence(this);
     }
 }
 
+class TowerDefence extends JFrame implements Runnable {
+    static final int FPS = 60;
+    private CashVault cashVault;
+    private Game game;
+    private Board board;
+    private List<Round> rounds;
+    private Player player;
+    private RoadSpawn roadSpawn;
 
+    private GameSpeed gameSpeed;
+
+
+    public TowerDefence(TowerDefenceBuilder builder) {
+
+        this.board = builder.board;
+        this.cashVault = builder.cashVault;
+        this.game = builder.game;
+        this.rounds = builder.rounds;
+        this.player = builder.player;
+        this.roadSpawn = builder.roadSpawn;
+        this.gameSpeed = builder.gameSpeed;
+    
+        setTitle("Tower Defence");
+        setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
+        setResizable(true);
+        setLocationRelativeTo(null);
+
+        // TODO: maybe not the best usage of shop stuff like this, change later
+        ShopModel shopModel = new ShopModel(player, cashVault,board);
+        ShopController shopController = new ShopController(shopModel);
+        
+        BoardView boardView = new BoardView(this.board, this.game, shopController);
+        add(boardView);
+        
+        pack();
+        setVisible(true);
+
+    }
+
+
+    @Override
+    public void run() {
+        for (Round round : rounds) {
+            while (!round.isRoundOver() || game.enemiesLeft() > 0) {
+                if (player.isAlive()) {
+                    Optional<EnemyFactory> spawn = round.spawn();
+
+                    if (spawn.isPresent()) {
+                        RoadToken token = new RoadToken(roadSpawn);;
+                        game.addEnemy(spawn.get().createEnemy(token));
+                    }
+
+                    List<Enemy> killed = game.update();
+                    for (Enemy enemy : killed) {
+                        cashVault.deposit(enemy.getReward());
+                    }
+                    
+                    board.update();
+                    repaint();
+                    try {
+                        Thread.sleep(1000 / gameSpeed.getSpeed());
+                    } catch (InterruptedException e) {
+                        e.printStackTrace();
+                    }
+                }
+            }
+        }
+        System.out.println("Game Over");
+    }
+
+}
 
 public class Main {
     public static void main (String[] args) throws Exception {
     
+    
         Point2D position = new Point2D.Double(100, 100);
         Tower tower = new LightningSmurfFactory().createTower(position);
         Tower tower2 = new LightningSmurfFactory().createTower(new Point2D.Double(400,300));
+        Board board = new Board(new Dimension(815, 635));
         Player player = new Player(30);
         CashVault cashVault = new CashVault(100);
-        Board board = new Board(new Dimension(815, 635));
-        Game game = new Game();       
+        Game game = new Game();
 
+        
         List<Round> rounds = RoundConfig.createRounds(cashVault);
+
 
         RoadSpawn spawn = new RoadBuilder(board, player)
                 .add(new Point2D.Double(500., 3.))
@@ -154,19 +228,17 @@ public class Main {
                 .add(new Point2D.Double(778., 303.))
                 .build();
 
-
         board.addTower(tower);
         board.addTower(tower2);
 
         TowerDefence towerDefence = new TowerDefenceBuilder()
+            .setBoard(board)
+            .setRounds(rounds)
+            .setPlayer(player)
             .setCashVault(cashVault)
             .setGame(game)
-            .setBoard(board)
-            .setPlayer(player)
             .setRoadSpawn(spawn)
-            .setRounds(rounds)
-            .setStartState(new MidRoundState())
-            .setGameSpeed(GameSpeed.SLOW)
+            .setGameSpeed(GameSpeed.NORMAL)
             .build();
 
         Thread thread = new Thread(towerDefence);
