+++ src/main/java/com/group34/Model/Game/Game.java
@@ -4,7 +4,6 @@ import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
 
-import com.group34.GameState.GameState;
 import com.group34.Model.Enemy.Enemy;
 
 public class Game {
@@ -20,14 +19,6 @@ public class Game {
         return enemies.iterator();
     }
 
-    public void removeEnemy(Enemy enemy) {
-        enemies.remove(enemy);
-    }
-
-    public TowerNotifier getNotifier() {
-        return notifier;
-    }
-  
     public int enemiesLeft() {
         return enemies.size();
     }
