+++ src/main/java/com/group34/View/Game/Game.java
@@ -28,4 +28,8 @@ public class Game extends JFrame {
         //setContentPane(container); // Set container as content
         //setVisible(true);
     }
+
+    public static void main(String[] args) {
+        SwingUtilities.invokeLater(() -> new Game());
+    }
 }
