+++ src/main/java/com/group34/View/Game/GameStatus.java
@@ -1,0 +1,5 @@
+package com.group34.View.Game;
+
+public enum GameStatus {
+    ACTIVE, PAUSED, GAME_OVER;
+}
